+++ example/session/src/main/java/org/apache/iotdb/doublewrite/DoubleWriteExample.java
@@ -1,4 +1,22 @@
-package org.apache.iotdb.session.doublewrite;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.iotdb.doublewrite;
 
 import org.apache.iotdb.rpc.IoTDBConnectionException;
 import org.apache.iotdb.rpc.StatementExecutionException;
@@ -8,7 +26,7 @@ import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DoubleWriteTest {
+public class DoubleWriteExample {
   private static Session session;
   private static String deviceId;
 
@@ -16,10 +34,8 @@ public class DoubleWriteTest {
       throws IoTDBConnectionException, StatementExecutionException {
     session = new Session("127.0.0.1", 6668, "root", "root");
     session.open();
-    // session.setStorageGroup("root.dw");
     deviceId = "root.dw.d01";
     testInsertRecords();
-    // session.deleteData("root.dw.d01", 233L);
     session.close();
   }
 
@@ -62,11 +78,13 @@ public class DoubleWriteTest {
         deviceIds.clear();
         measurementsList.clear();
         valuesList.clear();
+        typesList.clear();
         timestamps.clear();
       }
     }
 
     session.insertRecords(deviceIds, timestamps, measurementsList, typesList, valuesList);
-    System.out.println("Producer: " + (double) produceCnt / (double) produceTime * 1000000000.0 + "/s");
+    System.out.println(
+        "Producer: " + (double) produceCnt / (double) produceTime * 1000000000.0 + "/s");
   }
 }
