+++ example/session/src/main/java/org/apache/iotdb/doublewrite/MultiDoubleWriteExample.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.iotdb.doublewrite;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class MultiDoubleWriteExample {
+  private static ArrayList<SingleThread> threads = new ArrayList<>();
+  private static String host = "127.0.0.1";
+  private static int rpcPort = 6667;
+
+  public static void main(String[] args) {
+    for (int i = 0; i < 10; i++) {
+      String deviceId = "root.dw.d0" + (i / 2 + 1);
+      List<String> measurements = new ArrayList<>();
+      measurements.add("s1");
+      measurements.add("s2");
+      measurements.add("s3");
+      SingleThread threadA = new SingleThread(host, rpcPort, deviceId, measurements);
+      threads.add(threadA);
+
+      measurements = new ArrayList<>();
+      measurements.add("s4");
+      measurements.add("s5");
+      measurements.add("s6");
+      SingleThread threadB = new SingleThread(host, rpcPort, deviceId, measurements);
+      threads.add(threadB);
+    }
+
+    for (SingleThread thread : threads) {
+      new Thread(thread).start();
+    }
+  }
+}
