+++ example/session/src/main/java/org/apache/iotdb/doublewrite/SingleThread.java
@@ -1,4 +1,22 @@
-package org.apache.iotdb.session.doublewrite;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.iotdb.doublewrite;
 
 import org.apache.iotdb.rpc.IoTDBConnectionException;
 import org.apache.iotdb.rpc.StatementExecutionException;
@@ -55,6 +73,7 @@ public class SingleThread implements Runnable {
           produceTime += endTime - startTime;
           deviceIds.clear();
           measurementsList.clear();
+          typesList.clear();
           valuesList.clear();
           timestamps.clear();
         }
@@ -62,7 +81,8 @@ public class SingleThread implements Runnable {
 
       session.insertRecords(deviceIds, timestamps, measurementsList, typesList, valuesList);
       session.close();
-      System.out.println("Producer: " + (double) produceCnt / (double) produceTime * 1000000000.0 + "/s");
+      System.out.println(
+          "Producer: " + (double) produceCnt / (double) produceTime * 1000000000.0 + "/s");
     } catch (IoTDBConnectionException | StatementExecutionException e) {
       e.printStackTrace();
     }
