+++ server/src/main/java/org/apache/iotdb/db/doublewrite/DoubleWriteProducer.java
@@ -1,3 +1,21 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package org.apache.iotdb.db.doublewrite;
 
 import org.apache.iotdb.service.rpc.thrift.TSInsertRecordsReq;
@@ -7,7 +25,6 @@ import java.util.concurrent.BlockingQueue;
 
 public class DoubleWriteProducer {
   private BlockingQueue<Pair<DoubleWriteType, TSInsertRecordsReq>> doubleWriteQueue;
-  private int fullCnt = 0;
 
   public DoubleWriteProducer(
       BlockingQueue<Pair<DoubleWriteType, TSInsertRecordsReq>> doubleWriteQueue) {
@@ -16,16 +33,9 @@ public class DoubleWriteProducer {
 
   public void put(Pair<DoubleWriteType, TSInsertRecordsReq> reqPair) {
     try {
-      if (doubleWriteQueue.size() == 1024) {
-        ++fullCnt;
-      }
       doubleWriteQueue.put(reqPair);
     } catch (InterruptedException e) {
       e.printStackTrace();
     }
-  }
-
-  public int getFullCnt() {
-    return fullCnt;
   }
 }
